pyproject.toml
[tool.poetry]
name = "course-work-6-django"
version = "0.1.0"
description = ""
authors = ["mvladlena85 <mvladlena@yandex.ru>"]
readme = "README.md"
packages = [{include = "course_work_6_django"}]

[tool.poetry.dependencies]
python = "^3.10"
django-cors-headers = "^3.13.0"
Django = "^4.1.5"
drf-spectacular = "^0.25.1"
python-dotenv = "^0.21.1"
psycopg2 = "^2.9.5"
django-phonenumber-field = "^7.0.2"
django-templated-mail = "^1.1.1"
djangorestframework = "3.13.1"
djangorestframework-simplejwt = "4.8.0"
phonenumbers = "^8.13.5"
djoser = "^2.1.0"
pillow = "^9.4.0"
drf-nested-routers = "^0.93.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
